### 2021-05-13 #3280 @whedon
[PRE REVIEW]: ddop: A python package for data-driven operations management
**Submitting author:** @AndreasPhilippi (<a href="http://orcid.org/0000-0002-6508-9128">Andreas Philippi</a>)
**Repository:** <a href="https://github.com/opimwue/ddop" target ="_blank">https://github.com/opimwue/ddop</a>
**Version:** v0.6.5
