### 2021-05-13 #3280 @danielskatz
@AndreasPhilippi - Operations Management is two words in English - please change this in your paper.  In addition, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
