### 2021-06-08 #3280 @AndreasPhilippi
> Also, because the code size is a little on the small side, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.

@danielskatz - Do you have any news for us regarding the query scope?
