### 2021-06-08 #3280 @danielskatz
@AndreasPhilippi - I'm sorry for the delay, but this discussion has not been as brief as most other editor discussions - I'm pinging the AEiC currently on duty about this: @openjournals/joss-eics to help move it forward to a decision
