### 2021-06-15 #3280 @AndreasPhilippi
@kthyng @danielskatz - The goal of ddop is to allow previously developed data-driven approaches (from the field of operations management) to be efficiently applied on different datasets.
Let me give you an example: In recent years, many algorithms have been proposed to solve the newsvendor problem based on historical data (including past demand observations as well as additional features such as weather forecasts or calendar data.). Now ddop integrates a wide range of these algorithms making it possible to apply and evaluate them on different datasets with only a few lines of code. So ddop is basically the same as machine learning libraries like scikit-learn. It' s just much smaller and instead of ML algorithms, it provides data-driven decision algorithms from the field of operations management.

However, we will update the paper with the "statement of need section" as soon as possible.
