### 2021-06-15 #3280 @danielskatz
Thanks - adding some of these specific examples will be helpful
