### 2021-06-25 #3280 @kyleniemeyer
Hi @AndreasPhilippi, the JOSS editorial board has reviewed your submission and found that it is within our scope, so we will move forward with review.

However, due to low editor availability, this will enter our waitlist as we work through the backlog of submissions. Thanks for your patience!
