### 2021-06-28 #3280 @arfon
:wave: @rlucas7, @g4brielvs, @kramea – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration here is [_ddop: A python package for data-driven operations management_](https://github.com/opimwue/ddop)
