### 2021-06-28 #3280 @kramea
Sure, I can review that.

On Mon, Jun 28, 2021 at 1:46 AM Arfon Smith ***@***.***>
wrote:

> 👋 @rlucas7 <https://github.com/rlucas7>, @g4brielvs
> <https://github.com/g4brielvs>, @kramea <https://github.com/kramea> –
> would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The submission under consideration here is *ddop: A python package for
> data-driven operations management* <https://github.com/opimwue/ddop>
>
