### 2021-09-19 #3429 @kramea
Here is my review of the software:

**Installation:**

- I created an empty environment, installed the requirements and then installed the ddop package. Importing the ddop package worked.
- However, there were some issues when I tried to import ddop.newsvendor. It has a dependency of 'mpmath'. It needs to be included in the dependencies.

**Functionalities & Tutorials:**

- I tried to follow the newsvendor tutorial with Yaz dataset. I believe their Tutorial page is a bit outdated (which is understandable considering the ongoing development). A lot of the usage has changed from the tutorials, including the names of the model algorithms. For instance, I could not find DecisionTreeWeighted Newsvendor.
- Once I corrected the names, I was able to run the code to assess the functionality. It works as intended and claimed by the authors.
- Their API reference also includes descriptions of the commands.
- One suggestion would be to include more documentation on how to load custom dataset, as the authors claim this package can be used as a common framework for any kind of dataset.

Paper:

- The authors have done a good job detailing the need for the package for operations management, as well as future work and references.
