### 2021-09-22 #3429 @AndreasPhilippi
@kramea Thank you four your good feedback.
 I have added 'mpmath' to the dependencies. I also updated the tutorial on the documentation website, so everything should be up to date now. Please note that we are continuously trying to extend the documentation.
