### 2021-10-13 #3429 @g4brielvs
Here is my review (at last). Thousands apologies for my missing in action.

### General checks

- The packages passes general checks. However, some of the dependencies were not included, e.g. `mpmath`.

### Documentation & Paper

- As indicated on the checklist, both document and paper pass the inspection. However, it would be beneficial to have more descriptions and expanded examples on the documentation, especially for those not familiar with the research area.
